+++ src/cmd.rs
@@ -9,6 +9,8 @@ use tui::{MsgTarget, Timestamp};
 use utils;
 use serde::de::{Deserializer, Visitor};
 
+use notifier::Notifier;
+
 pub struct Cmd {
     /// Command name. E.g. if this is `"cmd"`, `/cmd ...` will call this command.
     pub name: &'static str,
@@ -125,7 +127,7 @@ impl<'de> Deserialize<'de> for AutoCmd {
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
-static CMDS: [&'static Cmd; 13] = [
+static CMDS: [&'static Cmd; 14] = [
     &AWAY_CMD,
     &CLEAR_CMD,
     &CLOSE_CMD,
@@ -137,6 +139,7 @@ static CMDS: [&'static Cmd; 13] = [
     &MSG_CMD,
     &NAMES_CMD,
     &NICK_CMD,
+    &NOTIFY_CMD,
     &RELOAD_CMD,
     &SWITCH_CMD,
 ];
@@ -569,6 +572,69 @@ fn switch(args: &str, _: &Poll, tiny: &mut Tiny, _: MsgSource) {
         );
     }
     tiny.tui.switch(words[0]);
+}
+
+////////////////////////////////////////////////////////////////////////////////////////////////////
+
+static NOTIFY_CMD: Cmd = Cmd {
+    name: "notify",
+    cmd_fn: notify,
+};
+
+fn notify(args: &str, _: &Poll, tiny: &mut Tiny, src: MsgSource) {
+    let words: Vec<&str> = args.split_whitespace().collect();
+
+    let mut show_usage = || {
+        tiny.tui.add_client_err_msg(
+            "/notify usage: /notify [off|mentions|messages]",
+            &MsgTarget::CurrentTab,
+        )
+    };
+
+    if words.len() == 0 {
+        tiny.tui.show_notify_mode(&MsgTarget::CurrentTab);
+    } else if words.len() != 1 {
+        return show_usage();
+    } else {
+        let notifier =
+            match words[0] {
+                "off" => {
+                    tiny.tui.add_client_notify_msg(
+                        "Notifications turned off",
+                        &MsgTarget::CurrentTab,
+                    );
+                    Notifier::Off
+                }
+                "mentions" => {
+                    tiny.tui.add_client_notify_msg(
+                        "Notifications enabled for mentions",
+                        &MsgTarget::CurrentTab,
+                    );
+                    Notifier::Mentions
+                }
+                "messages" => {
+                    tiny.tui.add_client_notify_msg(
+                        "Notifications enabled for all messages",
+                        &MsgTarget::CurrentTab,
+                    );
+                    Notifier::Messages
+                }
+                _ => {
+                    return show_usage();
+                }
+            };
+        // can't use `MsgSource::to_target` here, `Serv` case is different
+        let tab_target =
+            match src {
+                MsgSource::Serv { ref serv_name } =>
+                    MsgTarget::AllServTabs { serv_name },
+                MsgSource::Chan { ref serv_name, ref chan_name } =>
+                    MsgTarget::Chan { serv_name, chan_name },
+                MsgSource::User { ref serv_name, ref nick } =>
+                    MsgTarget::User { serv_name, nick },
+            };
+        tiny.tui.set_notifier(notifier, &tab_target);
+    }
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////