+++ src/lib.rs
@@ -7,6 +7,7 @@
 #![feature(entry_and_modify)]
 #![feature(global_allocator)]
 #![feature(inclusive_range_syntax)]
+#![feature(nll)]
 #![feature(offset_to)]
 
 extern crate alloc_system;
@@ -39,6 +40,7 @@ mod cmd;
 mod cmd_line_args;
 mod conn;
 mod logger;
+mod notifier;
 mod stream;
 mod wire;
 pub mod config;